DRAW-348 refactor: Request Dto 이관
