+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/dto/game/SessionInitializeRequest.kt
@@ -1,4 +1,4 @@
-package com.xorker.draw.websocket
+package com.xorker.draw.websocket.message.request.dto.game
 
 data class SessionInitializeRequest(
     val accessToken: String,
