+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/session/SessionFactory.kt
@@ -6,12 +6,12 @@ import com.xorker.draw.room.RoomId
 import com.xorker.draw.room.RoomRepository
 import com.xorker.draw.user.User
 import com.xorker.draw.user.UserId
-import com.xorker.draw.websocket.MafiaGameRandomMatchingRequest
 import com.xorker.draw.websocket.Session
-import com.xorker.draw.websocket.SessionInitializeRequest
 import com.xorker.draw.websocket.WaitingQueueSession
 import com.xorker.draw.websocket.message.request.SessionWrapper
 import com.xorker.draw.websocket.message.request.WaitingQueueSessionWrapper
+import com.xorker.draw.websocket.message.request.dto.game.MafiaGameRandomMatchingRequest
+import com.xorker.draw.websocket.message.request.dto.game.SessionInitializeRequest
 import org.slf4j.MDC
 import org.springframework.stereotype.Component
 import org.springframework.web.socket.WebSocketSession
